requirements.txt
duckdb>=0.9.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
weasyprint>=60.0
python-telegram-bot>=20.0
numpy>=1.25.0
